refactor(core): move PairingError into hid::pairing, point ipc.rs at core
PairingError is pure data (thiserror, String/u8/&'static str payloads,
no hidpp/async-hid types), so it belongs beside Click/PasskeyMethod/
ReceiverSelector in openlogi_core::hid::pairing rather than in
openlogi-hid. It is not itself a wire type — PairingFailure is, via
ipc.rs's `impl From<PairingError> for PairingFailure` — but preparing
this move keeps that conversion's source type in the same crate as
every other pure pairing type.

openlogi-agent-core/src/ipc.rs's `use openlogi_hid::{...}` named nine
types; eight were already re-exports of openlogi_core::hid, and the
ninth (PairingError) now is too, so the import points at
openlogi_core::hid directly and the IPC contract module stops naming
openlogi_hid at all.

`impl From<async_hid::HidError> for PairingError` hit the same orphan
rule as WriteError's did: replaced with a local classify_hid_error
function at its two call sites (enumerate_hidpp_devices /
open_hidpp_channel in list_pairing_receivers and open_receiver).

This is preparation, not the fix: openlogi-agent-core still depends on
openlogi-hid unconditionally for hardware.rs's agent-side orchestration,
so the build graph is unchanged. Wire goldens (wire_format.rs) pass
unchanged; PROTOCOL_VERSION stays at 16.
