+++ crates/openlogi-core/src/hid.rs
@@ -18,7 +18,7 @@ pub mod smartshift;
 pub use dpi::{DpiCapabilities, DpiInfo};
 pub use error::{HidppFeatureErrorKind, HidppOperation, WriteError};
 pub use light::{LightCommand, commands_for_light_settings};
-pub use pairing::{Click, PasskeyMethod, ReceiverSelector};
+pub use pairing::{Click, PairingError, PasskeyMethod, ReceiverSelector};
 pub use route::{
     BOLT_PIDS, DIRECT_DEVICE_INDEX, DeviceRoute, LIGHTSPEED_PIDS, UNIFYING_PIDS, is_receiver_pid,
     receiver_display_name, speaks_unifying_protocol,
